Добавлен прогноз на 5 дней через каждые 3 часа.

+++ app/src/main/java/com/warinator/app/weatherornot/model/Wind.java
@@ -1,4 +1,5 @@
 package com.warinator.app.weatherornot.model;
+
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
@@ -9,7 +10,7 @@ public class Wind {
     private float speed;
     @SerializedName("deg")
     @Expose
-    private int deg;
+    private float deg;
 
     public float getSpeed() {
         return speed;
@@ -19,11 +20,11 @@ public class Wind {
         this.speed = speed;
     }
 
-    public int getDeg() {
+    public float getDeg() {
         return deg;
     }
 
-    public void setDeg(int deg) {
+    public void setDeg(float deg) {
         this.deg = deg;
     }
